Refactor & tidy up burger menu
