+++ website/src/theme.js
@@ -10,6 +10,7 @@ const theme = {
     navbarText: '#f4d62e',
   },
   boxShadow: 'rgba(30, 30, 30, 0.15) 0px 2px 40px 0px',
+  overlay: 'rgba(0, 0, 0, 0.85)', // Aligned with lightbox
 }
 
 export default theme
